Update Task and User model to add Timestamps - Chapter13 - 1
Chapter 13 - 1st video - Added timestamps as an option to userSchema and
taskSchema.
Actually, taskSchema was still not created. The model was created
immediatly with the definitions for tasks.
We refactored the code, so now we create first a taskSchema. This Schema
accepts 2 arguments (objects), the first one being the one we used
everytime (definitions) and the second being the options object. Here we
added {timestamps: true}