Add completed Query to url - Chapter 13 - 3rd video
Added the req.query.completed to the get /tasks route. So now we can put
in the end of the URL the query with ? and completed set to false or
true.
As the req.query param is a string, we don't use the value directly, but
we avaliate it: If req.query.completed === 'true' (String) - means the result
avaliates to true (Boolean) else, is false. If nothing is added, we
return an Empty Object (Match was previously created as an Empty
Object!).
Also, we changed a bit the req.user.poppulate: Now instead of having
directly 'tasks', it has an object with path: 'tasks', and then our
match object!