Add sortBy query - Chapter 13 - 5th video
Added the sortyBy query field. In this case we created an object called
sort, which will take the first part of the query - Note that the query
will have two parts: the name (createdAt for example) and the order
(asc/desc). So the name is one key of the sort object, and the order is
the value. The value will be transformed to -1 for desc, and 1 for asc.
Sort is part of the options object!