Handling Express errors - Chapter 14 - 5th Video
Here we are taking care of errors when files are too large or wrong
format.
Instead of having a big html response for the error, we want json.
For that, in the POST route, we can add a 4th cb function! (after the
usual (req, res). The signature of this callback function is:
(error, req, res, next). Basically, is called when an error occurs, and
it looks like this:

(normal callback here...)

}, (error, req, res, next) => {
	res.status(400).send({error: error.message})
})