Add image to req.user.avatar - Chapter 14 - 6th video
In this part, we send the picture not to a place in the filesystem (with
the dest field) but to the req Object (we just delete the dest field).
Then we just save the req.user.avatar with the file that is in
req.file.buffer, and save the user as usual.

To delete the avatar, we just need to pass undefined to req.user.avatar
and save the user!