Create route to get the image on the browser
Here we added a route to get the image in the browser.
We just find the image by id, set the type of response to image/jpg and
send the user.avatar.

In this case, the endpoint was /users/:id/avatar. Not quite sure why we
didn't use the me in the place if :id..

Actually, I tried with /me instead of /:id  in the browser.. the problem
is authentication! I was not authenticated and in postman the image
doesn't load.. maybe is because of this..